[Code] move RepoState to models (#44976)
RepoState is used by both front end and back end code, so move it to models.
