+++ x-pack/legacy/plugins/code/public/components/admin_page/project_item.tsx
@@ -24,9 +24,9 @@ import React from 'react';
 import { connect } from 'react-redux';
 import { Link } from 'react-router-dom';
 import { i18n } from '@kbn/i18n';
-import { Repository, WorkerReservedProgress } from '../../../model';
+import { Repository, RepoState, WorkerReservedProgress } from '../../../model';
 import { deleteRepo, indexRepo, initRepoCommand } from '../../actions';
-import { RepoState, RepoStatus } from '../../actions/status';
+import { RepoStatus } from '../../actions/status';
 
 const stateColor = {
   [RepoState.CLONING]: 'secondary',
