+++ x-pack/legacy/plugins/code/public/components/main/content.tsx
@@ -38,9 +38,10 @@ import {
   SearchScope,
   WorkerReservedProgress,
   Repository,
+  RepoState,
 } from '../../../model';
 import { CommitInfo, ReferenceInfo } from '../../../model/commit';
-import { changeSearchScope, FetchFileResponse, RepoState, RepoStatus } from '../../actions';
+import { changeSearchScope, FetchFileResponse, RepoStatus } from '../../actions';
 import { MainRouteParams, PathTypes } from '../../common/types';
 import { RootState } from '../../reducers';
 import {
